requirements.txt
# Data Collection
ccxt==4.2.25
websockets==12.0
aiohttp==3.9.3

# Data Processing
pandas==2.2.0
numpy==1.26.3
ta==0.11.0

# Storage
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
sqlalchemy==2.0.25
greenlet==3.0.3

# Notifications
python-telegram-bot==20.7

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
schedule==1.2.0
pyyaml==6.0.1
psutil==5.9.8

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
aiosqlite==0.19.0
fakeredis==2.21.1
